+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Implementation of `FloatDiff` for arrays of size 0 to 32 (inclusive) where the
+  type allows it.
 
 ## [0.1.3] - 2020-04-07
 ### Added